Added binarytreefile
The beginning of the binary tree. There is one class Node that represents the structure of a node which holds the value of the node and the children to the node. The class Bintree is the binary tree which for now has the attribute root and the method add. Thereafter we have the beginning of the function addNode.