LCA works for DAGs !!
The LCA-function works for one node, a line, a binary tree and DAGs. If there are more than one LCA, all of the LCAs are returned.

I'm not totally sure if the counter for how long the path is will work in all cases. I will have to do more tests to be sure about that.